:construction: 우선적으로 새로침만 경고창을 띄우도록 수정
