+++ src/app/[locale]/(main)/(plan)/_components/NotificationBox.tsx
