+++ src/app/[locale]/(main)/(plan)/_components/ResizableContent.tsx
