+++ src/app/[locale]/(main)/(plan)/_components/UnsavedChangesWarning.tsx
@@ -12,8 +12,11 @@ export default function UnsavedChangesWarning({
 
   useEffect(() => {
     const handleBeforeUnload = (e: BeforeUnloadEvent) => {
-      e.preventDefault();
-      e.returnValue = ''; // Chrome에서 필요
+      // 새로고침 시 경고 메시지 표시
+      const message =
+        '변경사항이 저장되지 않을 수 있습니다. 정말로 나가시겠습니까?';
+      e.returnValue = message; // 브라우저에 따라 이 메시지가 표시될 수 있음
+      return message; // 일부 브라우저에서는 이 반환값을 사용
     };
 
     const handlePageShow = (e: PageTransitionEvent) => {
@@ -33,9 +36,14 @@ export default function UnsavedChangesWarning({
 
     sessionStorage.setItem('pageLoadTimestamp', currentTime);
 
+    // 이벤트 리스너 추가
+    window.addEventListener('beforeunload', handleBeforeUnload);
+    window.addEventListener('pageshow', handlePageShow);
+
     return () => {
       window.removeEventListener('beforeunload', handleBeforeUnload);
       window.removeEventListener('pageshow', handlePageShow);
+      sessionStorage.removeItem('pageLoadTimestamp');
     };
   }, [router]);
