+++ src/app/[locale]/(main)/(plan)/layout.tsx
@@ -1,5 +1,5 @@
 import {ReactNode} from 'react';
-import UnsavedChangesWarning from '@/app/[locale]/(main)/(trip)/_components/UnsavedChangesWarning';
+import UnsavedChangesWarning from '@/app/[locale]/(main)/(plan)/_components/UnsavedChangesWarning';
 
 export default function TripLayout({children}: {children: ReactNode}) {
   return (
