+++ src/app/[locale]/(main)/(plan)/place/add/page.tsx
