+++ src/app/[locale]/(main)/(plan)/ready/page.tsx
